datadog service: update to use supervisor approach
Starting things manually has been deprecated for many versions and
recently even removed for some things (datadogstatsd). Further, with
this approach we don't have to worry about starting things ourselves, we
just have to patch the supervisor config so that it knows what to use
and it should do the work for us.